Replaced the forgotten hard-coded image size and changed width and height types to uint32_t

+++ OpenNeighborhood/src/Elements/Element.h
@@ -17,16 +17,16 @@ public:
 
 	inline std::string GetLabel() const { return m_Label; }
 	inline std::string GetTextureName() const { return m_TextureName; }
-	inline float GetWidth() const { return m_Width; }
-	inline float GetHeight() const { return m_Height; }
+	inline uint32_t GetWidth() const { return m_Width; }
+	inline uint32_t GetHeight() const { return m_Height; }
 
 	using EventCallbackFn = std::function<void(Event&)>;
 	inline void SetEventCallback(const EventCallbackFn& callback) { m_EventCallback = callback; }
 protected:
 	std::string m_Label;
 	std::string m_TextureName;
-	float m_Width;
-	float m_Height;
+	uint32_t m_Width;
+	uint32_t m_Height;
 	float m_Padding = 10.0f;
 	EventCallbackFn m_EventCallback;
